feat: Update content and CV links
This commit updates the content of the "about me" and "welcome" sections in both English and Spanish, reflecting the latest professional experience and skills. It also updates the links to the CVs in both languages. Additionally, it removes SQLite from the backend technologies section and adjusts the z-index of the header and footer components. Finally, it removes some of the old experiences.