feat: Add GitHub Finder project and update related components
This commit introduces a new project called "GitHub Finder," an application developed in Swift 6 and SwiftUI, to the project listings in `src/data/proyectos.tsx`. It updates the corresponding locale files (`es/proyectos.json` and `en/proyectos.json`) with project details. Additionally, the `PortfolioBox.tsx` component is modified to include "GitHub Finder" in the project display logic, and the `OtrasTecnologiasSectionAbout.tsx` component is updated to reflect the Swift technology icon.